requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
bcrypt==4.0.1
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0
slowapi==0.1.9
pybreaker==1.0.1
cryptography==41.0.7
psutil==5.9.6
websockets==12.0
redis==5.0.1
python-dotenv==1.0.0
requests==2.31.0 
aiofiles==23.2.1
aioredis==2.0.1
openai==1.12.0
google-generativeai==0.3.1
# AI and Computer Vision dependencies
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
pytesseract==0.3.10
face-recognition==1.3.0
scikit-learn==1.3.2
# Azure AI services
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-face==0.4.1
azure-ai-formrecognizer==3.3.0
azure-ai-openai==1.0.0
# Google AI services
google-cloud-vision==3.4.4
google-cloud-language==2.11.1
# AWS AI services
boto3==1.34.0
# Hugging Face
transformers==4.35.2
torch==2.1.1